refactor(waf): split bypass.rs into types / strategy / mutate
Pure structural move (no behavior change): the 1370-line waf/bypass.rs is
now waf/bypass/ with one concern per file —
- types.rs    mutation vocabulary + strategy/stats shapes
- strategy.rs per-WAF strategy selection + multi-WAF merge
- mutate.rs   the payload-rewrite implementations
- mod.rs      dispatch + the small re-exported surface

Callers keep using `crate::waf::bypass::{MutationType, MutationStats,
BypassStrategy, merge_strategies, apply_mutations_tagged}` unchanged. The 16
dispatch entry points plus the primitives the tests exercise are `pub(super)`
in mutate.rs; tests.rs reaches them via `use super::*` and is untouched. All
74 bypass tests pass, clippy clean.
